improved unit tests - sorting
